requirements.txt
langchain_community
SQLAlchemy
psycopg2-binary
python-dotenv
python-multipart
pydantic
python-flask
postgresql
